+++ model/nr-eesm-ir.h
@@ -85,7 +85,7 @@ protected:
    * \param mcs the MCS of the transmission
    * \return The equivalent MCS after retransmissions
    */
-  double GetMcsEq (uint16_t mcsTx) const override;
+  double GetMcsEq (uint8_t mcsTx) const override;
 
 private:
   double m_Reff {0.0};  //!< equivalent effective code rate after retransmissions
